Started working on the serverside.
I added a separate package for exceptions and another one for other utilities.
In the utilities package I created a thread safe class to be used for representing the On/Off state of a service.

Althought the UDP connection is now implemented, the Server uses only TCP currently.
The functionality for now consists of accepting players and reading their input successfully.